find_double_title/requirements.txt
contentful>=1.12.0
python-dotenv>=0.19.0


